pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wildfire-risk-eu"
version = "0.1.0"
description = "Asset-level wildfire exposure intelligence for Mediterranean WUI"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "geopandas>=0.14",
    "fiona>=1.9",
    "rasterio>=1.3",
    "rioxarray>=0.15",
    "xarray>=2024.1",
    "duckdb>=0.10",
    "numpy>=1.26",
    "pandas>=2.2",
    "scipy>=1.12",
    "matplotlib>=3.8",
    "folium>=0.16",
    "pyyaml>=6.0",
    "pyogrio>=0.7",
    "netcdf4>=1.6",
    "pyproj>=3.6",
    "cdsapi>=0.6",
    "requests>=2.31",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "ruff>=0.3",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-dir]
"" = "."

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
